pyproject.toml
[tool.poetry]
name = "currency-icons"
version = "1.0.0"
description = ""
authors = ["Albert <github@albert.sh>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.0.5"
lxml = "^4.5.2"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
